minor style fixes

+++ src/main/java/me/white/wlua/IntegerValue.java
@@ -18,7 +18,7 @@ public final class IntegerValue extends NumberValue {
     }
 
     @Override
-    public long toInteger() {
+    public int toInteger() {
         return value;
     }
