+++ src/main/java/me/white/wlua/StringValue.java
@@ -3,7 +3,6 @@ package me.white.wlua;
 import java.lang.foreign.Arena;
 import java.util.Objects;
 
-// TODO store byte[] value as well
 public final class StringValue extends LuaValue {
     private final String value;
 
@@ -40,8 +39,8 @@ public final class StringValue extends LuaValue {
     }
 
     @Override
-    public long toInteger() {
-        return (long)Double.parseDouble(value);
+    public int toInteger() {
+        return (int)Double.parseDouble(value);
     }
 
     @Override
